Moved sample test plugin to own directory.
